requirements.txt
flask
flask_sqlalchemy
sqlalchemy
psycopg2-binary
python-dotenv
gunicorn